refactor(devtools): improve events page component structure
